src/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
python-dotenv==1.1.1
pydantic-settings==2.10.1
aiofiles==24.1.0
langchain==0.3.27
langchain-community==0.3.27
PyMuPDF==1.26.3
motor==3.7.1
pydantic-mongo==3.1.0
qdrant-client==1.15.1
openai==1.99.1
ollama==0.5.2
celery==5.5.3
redis==6.4.0
kombu==5.5.4
billiard==4.2.1
vine==5.1.0